Split Term subclasses into their own files.
