+++ lib/term/iri.js
@@ -1,0 +1,5 @@
+/**
+ * "Symlink" to the real IRI module, but uses Node's require mechanism
+ * instead.
+ */
+module.exports = require('../iri');
